refactor(core): group document metadata helpers
Move attachments, destinations, layers, outline, structure, and viewer-state helpers under a document domain folder so the core root exposes fewer internal implementation files.
